certificate-publisher: Don't publish extra certs
Publish the default certificate and key of only whichever ingresscontroller
has the cluster ingress domain.

Before this commit, the certificate-publisher controller published the
certificates and keys of all ingresscontrollers, in the "router-certs"
secret.  However, the only component that uses the router-certs secret is
the authentication operator, which only needs the certificate and key for
the cluster ingress domain.  Moreover, collecting the certificates and keys
for all ingresscontrollers can produce a result that exceeds the maximum
secret size of 1 mebibyte, causing the certificate-publisher controller to
fail to create or update the router-certs secret.  This commit changes the
certificate-publisher controller not to publish the extraneous certificates
and keys.

This commit fixes OCPBUGS-853.

https://issues.redhat.com/browse/OCPBUGS-853

* pkg/operator/controller/certificate-publisher/controller.go: Update
comments to reflect that the controller only publishes the certificate and
key of the ingresscontroller for the cluster ingress domain in the
"router-certs" secret.
(New): Add a new predicate to the watch on ingresscontrollers, using the
new hasClusterIngressDomain method and isDefaultIngressController function.
(secretToIngressController): Skip ingresscontrollers for domains other than
the cluster ingress domain.
(hasClusterIngressDomain): New method.  Get the cluster ingress config and
return true iff the given ingresscontroller has the same domain.
(isDefaultIngressController): New function.  Return true iff the given
ingresscontroller is the "default" ingresscontroller.
(Reconcile): Get the cluster ingress config, and pass it to
ensureRouterCertsGlobalSecret.
* pkg/operator/controller/certificate-publisher/publish_certs.go
(ensureRouterCertsGlobalSecret): Add a parameter for the ingress config,
and pass the domain from the ingress config to
desiredRouterCertsGlobalSecret.
(desiredRouterCertsGlobalSecret): Add a parameter for the cluster ingress
domain, use it to filter out extraneous ingresscontrollers, and publish the
certificate and key for only the ingresscontroller that has the cluster
ingress domain.
(getDefaultCertificateSecretForIngressController): New function used by
desiredRouterCertsGlobalSecret.
* pkg/operator/controller/certificate-publisher/publish_certs_test.go
(newSecret): Delete the example PEM data and put the secret's name in the
"tls.crt" and "tls.key" data fields of the returned secret so that the data
fields differ for different secrets.
(TestDesiredRouterCertsGlobalSecret): Modify test cases so that they verify
that the router-cert secret's data fields have the expected values.  Add a
"default certificate, explicit" test case.  Add "custom certificate" and
"missing custom certificate" test cases for the default ingresscontroller
with a custom default certificate.  Add a "custom ingresscontroller with
the cluster ingress domain" test case.  Change the "no secrets" test case
to use the default ingresscontroller so that the test case would expect a
router-cert secret but for the missing secret for the ingresscontroller.
Update the "missing secret", "extra secret", and "perfect match" test cases
not to expect the router-cert secret to include data for other
ingresscontrollers.
* test/e2e/all_test.go (TestAll): Update the lists of tests.
* test/e2e/certificate_publisher_test.go: Delete file.  This deletes the
TestCreateIngressControllerThenSecret and
TestCreateSecretThenIngressController tests.
* test/e2e/operator_test.go (TestUpdateDefaultIngressController): Rename...
(TestUpdateDefaultIngressControllerSecret): ...to this.  Expand the test to
verify that the operator updates both the router-certs secret as well as
the default-ingress-cert configmap correctly, as well as verifying that the
operator *does not* update the router-certs secret and the configmap if the
ingresscontroller is updated to reference a non-existent secret.  Update
polling loops to use the same timeout value for consistency.