Move "default-ingress-cert" configmap publishing
Move the reconciliation of the "default-ingress-cert" configmap from the
"certificate" controller to the "certificate-publisher" controller in order
to ensure that the configmap gets updated when the default-certificate
secret is updated.

Before this commit, the certificate-publisher controller already published
the "router-certs" secret in the "openshift-config-managed" namespace.  The
router-certs secret is supposed to have the certificate and key of the
ingresscontroller that has the cluster ingress domain, so the controller
watches secrets as well as ingresscontrollers in order to update the
router-certs secret when the ingresscontroller is updated to reference a
different secret or when the content of the referenced secret is updated.

In contrast, the certificate controller was originally written to generate
a self-signed CA and generate default certificates for ingresscontrollers
using this CA, and for these purposes, the controller only needs to watch
ingresscontrollers and not the secrets themselves.

Commit 9640767788bd12843e4b12884e3e4e6727d3c810 added reconciliation of the
default-ingress-cert configmap to the certificate controller.  This
configmap has the default certificate of the "default" ingresscontroller.
The configmap should be updated when the secret reference or content is
updated.  Moreover, it makes sense conceptually for the
certificate-publisher controller to handle both publication tasks: both the
router-certs secret as well as the default-ingress-certs configmap.

* pkg/operator/controller/certificate/controller.go (Reconcile): Move the
call to ensureDefaultIngressCertConfigMap from here...
* pkg/operator/controller/certificate-publisher/controller.go (Reconcile):
...to here.
* pkg/operator/controller/certificate/publish_ca.go: Rename...
* pkg/operator/controller/certificate-publisher/publish_ca.go: ...to this.