Enable relative resource names for gcp dns zones
Enable the gcp dns provider to use relative resource names for the
managedZones. This allows the project to be specified as part of the
zone ID and override the defalut cluster project when managing the dns
records.

In addition to accepting the zone name as before, the zone ID
can now accept the following format:

projects/project-id/managedZones/zone-name