Absorb PodsScheduled condition into MinAvailable
Remove the PodsScheduled status condition on IngressController objects, and
report the information that PodsScheduled reported instead in the
DeploymentReplicasMinAvailable status condition's message when its status
is "False".

Before this commit, the PodsScheduled status condition could erroneously
signal that there was a problem when some misscheduled pods existed even
if there were enough correctly scheduled and available pods.

The PodsScheduled status condition was added in order to provide diagnostic
information when there was a problem.  That is, PodsScheduled was intended
to indicate *why* there was a problem, not *whether* there was a problem.
The DeploymentReplicasMinAvailable status condition definitively
indicates *whether* there is a problem, and its status message should
indicate *why* there is a problem.  Having a separate status condition for
the "why" is superfluous and unnecessarily complicated in this instance.

Follow-up to commit c5e0f4a513a327a42fb4a184ca1b1b9b7f39f505.

This bug fixes OCPBUGS-434.

https://issues.redhat.com/browse/OCPBUGS-434

* pkg/operator/controller/ingress/controller.go
(IngressControllerPodsScheduledConditionType): Delete const.
* pkg/operator/controller/ingress/metrics_test.go
(TestDeleteIngressControllerConditionsMetric): Delete a reference to the
"PodsScheduled" status condition in a test case.
* pkg/operator/controller/ingress/status.go (syncIngressControllerStatus):
Delete the call to computeDeploymentPodsScheduledCondition.  Pass the list
of pods to computeDeploymentReplicasMinAvailableCondition.
(PruneConditions): Delete pruning of the "DeploymentDegraded" status
condition (which was removed in 4.6.0) and instead prune the
"PodsScheduled" status condition.
(computeDeploymentPodsScheduledCondition): Rename...
(checkPodsScheduledForDeployment): ...to this.  Change the return value
from an operator status condition to an error value.
(computeDeploymentReplicasMinAvailableCondition): Call
checkPodsScheduledForDeployment and include the error string in the status
condition's message if the function returns an error.
(computeIngressDegradedCondition): Remove
IngressControllerPodsScheduledConditionType from the list of conditions to
check.
* pkg/operator/controller/ingress/status_test.go
(TestComputePodsScheduledCondition): Rename...
(Test_checkPodsScheduledForDeployment): ...to this.  Update for the
function's new name and result value.
(TestComputeIngressDegradedCondition): Delete test cases for PodsScheduled.
(TestComputeDeploymentReplicasMinAvailableCondition): Update to set the
deployment's label selector and pass in a list of pods (which is empty).