Test_desiredLoadBalancerService: Fix NodePort case
Don't check any service annotations if the service doesn't exist.

This fixes a logic error, which was obscured by the facts that (a) the test
sets ic.Status.EndpointPublishingStrategy.LoadBalancer to an empty value
rather than to nil even if the endpoint publishing strategy type is
"NodePortService" and (b) the only test case that specifies the
"NodePortService" endpoint publishing strategy type also specifies the AWS
platform type.  Specifying "NodePortService" on AWS worked because the
empty value for ic.Status.EndpointPublishingStrategy.LoadBalancer is
interpreted to indicate the "LoadBalancerService" strategy type with the
default "External" scope, and so the logic for AWS skipped the check for
the internal LB annotation on the non-existent service anyway, and the AWS
logic checks the endpoint publishing strategy type before checking other
annotations.  The logic to check annotations for other platforms does not
check the endpoint publishing strategy type at all.

* pkg/operator/controller/ingress/load_balancer_service_test.go
(Test_desiredLoadBalancerService): Don't check service annotations if the
service doesn't exist.