Test_desiredLoadBalancerService: Simplify logic
Simplify the logic of Test_desiredLoadBalancerService by making the
expectations of which annotations are set and which aren't into inputs to
the test cases and replacing individual checks for specific annotations
with a generic loop to verify these expectations.

* pkg/operator/controller/ingress/load_balancer_service_test.go
(Test_desiredLoadBalancerService): Add expectedServiceAnnotations input and
delete checks for specific annotations.
(checkServiceHasAnnotation): Delete helper function.