ensureWildcardDNSRecord: Delete ic parameter
Replace the operatorv1.IngressController parameter of
ensureWildcardDNSRecord with parameters for the DNSRecord name, labels,
owner reference, domain, and endpoint publishing strategy, and similarly
for currentWildcardDNSRecord, desiredWildcardDNSRecord, and
deleteWildcardDNSRecord.

* pkg/operator/controller/ingress/controller.go (ensureIngressDeleted):
Update call to deleteWildcardDNSRecord with the appropriate arguments.
(ensureIngressController): Update to call ensureWildcardDNSRecord with the
appropriate arguments.
* pkg/operator/controller/ingress/dns.go (ensureWildcardDNSRecord): Replace
the ic parameter with name, dnsRecordLabels, ownerRef, domain, and
endpointPublishingStrategy parameters.  Update calls to
desiredWildcardDNSRecord and currentWildcardDNSRecord with the appropriate
arguments.
(desiredWildcardDNSRecord): Replace the ic parameter with name,
dnsRecordLabels, ownerRef, dnsRecordDomain, and endpointPublishingStrategy
parameters.
(currentWildcardDNSRecord, deleteWildcardDNSRecord): Replace the ic
parameter with a name parameter.
* pkg/operator/controller/ingress/dns_test.go
(TestDesiredWildcardDNSRecord): Update call to desiredWildcardDNSRecord.