ensureInternalIngressControllerService: Updates
Add logic to ensureInternalIngressControllerService to update the service
when an update is required.

* pkg/operator/controller/ingress/internal_service.go
(ensureInternalIngressControllerService): Use the new updateInternalService
method to update the service as needed.
(updateInternalService): Check whether the given ClusterIP service needs to
be updated, and update it if so, using the new internalServiceChanged
function.
(managedInternalServiceAnnotations): New variable with the set of
annotation keys for annotations that the operator manages for its internal
router services.
(internalServiceChanged): New function.  Check whether the current internal
service needs to be updated, and update it if it does, using the new
managedInternalServiceAnnotations variable.
* pkg/operator/controller/ingress/internal_service_test.go: New file.
(Test_desiredInternalIngressControllerService): New test.  Verify that
desiredInternalIngressControllerService returns the expected service.
(Test_internalServiceChanged): New test.  Verify that
internalServiceChanged correctly detects changes and performs updates.