dns/aws: Store target hosted zone id in annotation
When upserting a DNS record in Route 53, store the target hosted zone id of
the ELB on the associated DNSRecord CR.  If a subsequent lookup of this id
fails for whatever reason, fall back to using the annotation value.

This change is necessary for Gateway API because Istio sometimes deletes
the ELB before the ingress operator deletes the DNS record, and so the ELB
no longer exists when the DNS record deletion logic needs to look up the
target hosted zone id of the ELB.

* pkg/dns/aws/dns.go (targetHostedZoneIdAnnotationKey): New const.
(Config): Add Client field.
(change): Use config.Client and targetHostedZoneIdAnnotationKey to annotate
the DNSRecord CR with the target hosted zone id of the ELB on upsert and
look up the id on deletion.
* pkg/operator/controller/dns/controller.go (Reconcile): Get the DNSRecord
CR before updating its status in case it has changed since the previous
get.
(createDNSProvider): Provide the client to the AWS DNS provider
implementation.