Avoid spurious updates for internalTrafficPolicy
Specify spec.internalTrafficPolicy on NodePort- and ClusterIP-type services
that the operator manages.  Also, ignore updates to the spec.ipFamilies and
spec.ipFamilyPolicy fields.

Before this commit, the update logic for NodePort- and ClusterIP-type
services would try to revert the default values that the API set for these
fields.

This commit fixes OCPBUGS-13190.

https://issues.redhat.com/browse/OCPBUGS-13190

* assets/router/service-cloud.yaml:
* assets/router/service-internal.yaml: Specify
"internalTrafficPolicy: Cluster".
* pkg/manifests/bindata.go: Regenerate.
* pkg/operator/controller/ingress/internal_service.go
(internalServiceChanged): Ignore spec.ipFamilies and spec.ipFamilyPolicy.
* pkg/operator/controller/ingress/internal_service_test.go
(Test_desiredInternalIngressControllerService): Verify that
spec.internalServiceChanged is set to "Cluster".
(Test_internalServiceChanged): Verify that changes to
spec.internalTrafficPolicy are detected and that changes to spec.ipFamilies
and spec.ipFamilyPolicy are ignored.
* pkg/operator/controller/ingress/nodeport_service.go
(desiredNodePortService): Set spec.internalTrafficPolicy to "Cluster".
(nodePortServiceChanged): Ignore spec.ipFamilies and spec.ipFamilyPolicy.
* pkg/operator/controller/ingress/nodeport_service_test.go
(TestDesiredNodePortService): Verify that spec.internalTrafficPolicy is set
to "Cluster".
(TestNodePortServiceChanged): Verify that changes to
spec.internalTrafficPolicy are detected and that changes to spec.ipFamilies
and spec.ipFamilyPolicy are ignored.