Add support for AWS shared VPC in another account
Add support for configuring DNS records in AWS Route 53 using a separate
account for the private hosted zone.

This commit resolves NE-1294.

https://issues.redhat.com/browse/NE-1294

* manifests/00-ingress-credentials-request.yaml: Add "sts:AssumeRole" to
the CredentialsRequest for AWS.
* pkg/manifests/bindata.go: Regenerate.
* pkg/dns/aws/dns.go (Config): Add a RoleARN field.
(NewProvider): If config.RoleARN is set, use it to configure the AWS client
using the specified role.
* pkg/dns/split/dns.go: New file.  Define a DNS provider implementation
that wraps two other DNS providers, using one of them to publish records to
the public zone and the other to publish records to the private zone.
(Provider): New type.  Store the private and public DNS providers, as well
as the private zone so that the Ensure, Delete, and Replace methods can use
it to determine whether they are publishing to the public zone or to the
private zone.
(NewProvider): New function.  Return a split DNS provider.
(Ensure, Delete, Replace): New methods.  Implement the dns.Provider
interface by calling the respective methods on the wrapped private and
public DNS providers.
* pkg/dns/split/dns_test.go (TestSplitDNSProvider): Verify that the split
DNS provider correctly dispatches to the private or public DNS provider as
appropriate, using fakeProvider.
(fakeProvider): New type.  Define a fake named DNS provider that records
its name when invoked.
(Ensure, Delete, Replace): New methods for fakeProvider to record
invocations and implement the dns.Provider interface.
(newFakeProvider): New function.  Return a fake provider.
* pkg/operator/controller/dns/controller.go (Config): Add a
"PrivateHostedZoneAWSEnabled" field to indicate whether the
"PrivateHostedZoneAWS" feature gate is enabled.
(createDNSProvider): Use the new split DNS provider and the AWS DNS
provider's new RoleARN configuration option to configure separate DNS
providers for public and private zones when a role ARN for the private zone
is specified in the cluster infrastructure config if the
"PrivateHostedZoneAWS" feature gate is enabled.
* pkg/operator/operator.go (New): Check the "PrivateHostedZoneAWS" feature
gate and specify it in the DNS controller config.