status: Add Gateway API objects to relatedObjects
This commit resolves NE-1277.

https://issues.redhat.com/browse/NE-1277

* pkg/operator/controller/names.go (GlobalOperatorsNamespace): New const
for the "openshift-operators" namespace.
(ServiceMeshSubscriptionName): Use GlobalOperatorsNamespace.
* pkg/operator/controller/status/controller.go (New): Add a watch for
subscriptions so that the status controller updates relatedObjects when the
OSSM subscription is added or removed.
(Config): Add GatewayAPIEnabled field.
(Reconcile): If GatewayAPIEnabled is true, add resources related to Gateway
API to relatedObjects.  If haveOSSMSubscription is true, add resources that
require the subscription.
(operatorState): Add haveOSSMSubscription field.
(getOperatorState): Set haveOSSMSubscription.
* pkg/operator/operator.go (New): Watch the "openshift-operators"
namespace, and specify GatewayAPIEnabled in the status controller's config.