status: Add Gateway API objects to relatedObjects
This commit resolves NE-1277.

https://issues.redhat.com/browse/NE-1277

* pkg/operator/controller/status/controller.go (New): Add a watch for
subscriptions so that the status controller updates relatedObjects with the
OSSM subscription when it is added or removed.
(Config): Add GatewayAPIEnabled field.
(Reconcile): Add dnsrecords in the operand namespace to relatedObjects.
If GatewayAPIEnabled is true, add resources related to Gateway API.
If haveOSSMSubscription is true, add resources that require the subscription.
(operatorState): Add haveOSSMSubscription field.
(getOperatorState): Set haveOSSMSubscription.
* pkg/operator/operator.go (New): Specify GatewayAPIEnabled in the status
controller's config.
* test/e2e/operator_test.go (TestClusterOperatorStatusRelatedObjects): Expect to
observe "dnsrecords" for the "openshift-ingress" namespace as well as
"gatewayclasses" in relatedObjects.