canary: Tolerate infra node NoExecute taint
Configure the canary daemonset to tolerate a
"node-role.kubernetes.io/infra" node taint that specifies the "NoExecute"
effect (or any other effect).

Before this commit, the canary daemonset tolerated the
"node-role.kubernetes.io/infra" node taint only if it specified the
"NoSchedule" effect.  However, some cluster admins taint infra nodes using
both "NoSchedule" as well as "NoExecute", in which case canary pods would
not be scheduled on these nodes.  This inconsistency caused some confusion.
With this commit, canary pods should be scheduled on infra nodes whether or
not they are tainted using the "NoExecute" effect.

This commit fixes OCPBUGS-9274.

https://issues.redhat.com/browse/OCPBUGS-9274

* assets/canary/daemonset.yaml: Tolerate the
"node-role.kubernetes.io/infra" taint irrespective of the specified effect.
* pkg/manifests/bindata.go: Regenerate.
* pkg/operator/controller/canary/daemonset_test.go
(TestDesiredCanaryDaemonSet): Expect "Effect" to be unspecified on the
toleration.
(TestCanaryDaemonsetChanged): Add a test case to verify that
canaryDaemonSetChanged correctly detects the change and updates the
daemonset if the specified effect changes on the toleration.