desiredRouterDeployment: Set HostPort if needed
Set spec.template.spec.containers[*].ports[*].hostPort on a router
deployment if the deployment sets spec.template.spec.hostNetwork to true.

Before this commit, the operator left the hostPort field unspecified. As a
result, the API set a default value for hostPort.  The operator detected
this as an external update and attempted to revert it.  This commit
prevents the operator from making these spurious updates in response to API
defaulting.

This commit fixes OCPBUGS-14995.

https://issues.redhat.com/browse/OCPBUGS-14995

* pkg/operator/controller/ingress/deployment.go (desiredRouterDeployment):
Set each container port's HostPort to the port's ContainerPort value when
HostNetwork is enabled for the deployment.
* pkg/operator/controller/ingress/deployment_test.go (checkContainerPort):
Verify that HostPort is set to ContainerPort if HostNetwork is enabled.