Don't add clientca-configmap finalizer if deleting
Modified-by: candita

* pkg/operator/controller/clientca-configmap/controller_test.go
In this backport to 4.13, we replace time.DateTime with time.RFC3339, since
time.DateTime is not available in go 1.19.

Don't add the clientca-configmap finalizer to an IngressController if it is
marked for deletion.

Before this commit, the clientca-configmap controller would attempt to
add (or re-add) its finalizer to an IngressController that had been marked
for deletion.  This attempt resulted in the following error from the API:

    IngressController.operator.openshift.io "test-client-ca-configmap" is invalid: metadata.finalizers: Forbidden: no new finalizers can be added if the object is being deleted, found new finalizers []string{"ingresscontroller.operator.openshift.io/finalizer-clientca-configmap"}

The controller would retry adding the finalizer repeatedly, causing the
same error to be repeatedly logged until the deletion completed.

This commit fixes OCPBUGS-14994.

https://issues.redhat.com/browse/OCPBUGS-14994

* pkg/operator/controller/clientca-configmap/controller.go (Reconcile):
Only attempt to add the finalizer if the IngressController is not marked
for deletion.
* pkg/operator/controller/clientca-configmap/controller_test.go: New file.
(Test_Reconcile): New test.  Verify that Reconcile adds and removes the
finalizer and creates, updates, or deletes the target configmap as
appropriate.
(fakeClientRecorder): New type, used in Test_Reconcile.
(Create, Delete, Update): New methods for fakeClientRecorder to implement
the controller-runtime client.Client interface.