If mTLS is required and the canary receives a "certificate required" TLS error, consider the canary probe successful.
Without an API to provide the canary with a valid client certificate,
the canary probes are guaranteed to fail when mTLS is required by the
default ingress controller.

However, if the failure is determined to be because the canary is
missing a client certificate, it's reasonable to assume that the ingress
controller is functional. This change checks specifically for the TLS
alert "certificate required", and if the default ingress controller also
has mTLS required, the probe function returns nil, indicating a
successful probe.