Check public DNS zone when reporting status
When reporting the "DNSReady" status condition on an IngressController,
check the status conditions for both the public zone and the private zone
of the associated DNSRecord.

Before this commit, only the status condition for the private zone was used
to compute the "DNSReady" status condition.  If the operator failed to
publish a DNS record in the public zone, this failure was not reported in
the IngressController's status conditions or the "ingress" clusteroperator
status conditions.

Follow-up to commit 9f329233f31db93d7e7fa6061e1f514dc02af5dc.

This commit fixes OCPBUGS-15978.

https://issues.redhat.com/browse/OCPBUGS-15978

* pkg/operator/controller/ingress/status.go (checkZoneInConfig): Use the
new zonesMatch helper function to check both the public zone as well as the
private zone.
(zonesMatch): New function.  Return a Boolean value indicating whether two
DNS zones match, based on their respective ID or "Name" tags.
* pkg/operator/controller/ingress/status_test.go (Test_checkZoneInConfig):
Verify that checkZoneInConfig checks the public zone as well as the private
zone.  Delete "[PrivateZone]" from test-case descriptions as each test case
is now tested as both a public zone and a private zone.