gateway-service-dns: Use gateway addresses
* pkg/operator/controller/gateway-service-dns/controller.go
(managedByIstioLabelKey): Remove const, which is no longer needed after
subsequent changes.
(NewUnmanaged): Remove the watch on services and the watch's associated
predicates.  Enqueue reconciliation requests for gateways instead of
services.  Define a predicate that checks the gateway addresses, using the
new gatewayAddressesChanged helper function, and use this predicate to
enqueue a reconciliation request for a gateway when any of its addresses
changes.
(gatewayAddressesChanged): New function.  Return a Boolean value indicating
whether the given gateway's addresses have changed.
(Reconcile): Expect the reconciliation request to be for a gateway instead
of a service.  Use the new getGatewayAddresses helper function to get the
gateway's addresses, and pass these addresses to ensureDNSRecordsForGateway
and deleteStaleDNSRecordsForGateway.
(getGatewayAddresses): New helper function.  Return a list of addresses
that can be used as targets for a DNSRecord CR, as well as a record type.
(ensureDNSRecordsForGateway): Replace the service parameter with "targets"
and "recordType" parameters.  Pass targets and recordType to
EnsureDNSRecord.  Change the owner reference for the DNSRecord CR to
specify the gateway instead of the service.
(deleteStaleDNSRecordsForGateway): Replace the service parameter with a
"targets" parameter.  Use the new targets parameter and existing domains
parameter to preserve any DNSRecord CR that has both a valid domain and a
valid target.
* pkg/operator/controller/gateway-service-dns/controller_test.go
(Test_Reconcile): Add addresses to gateway test inputs, and remove services
as test inputs.  Add a test case for a gateway with an address but no
listeners, as well as a test case for various combinations of addresses.
(Test_gatewayAddressesChanged): New test.  Verify that
gatewayAddressesChanged behaves as expected.
* pkg/resources/dnsrecord/dns.go (EnsureDNSRecord): Replace the service
parameter with "targets" and "recordType" parameters.  Pass these new
parameters to desiredDNSRecord.
(desiredDNSRecord): Replace the service parameter with "targets" and
"recordType" parameters.  Move service-specific logic from here...
(desiredWildcardDNSRecord): ...to here.