[release-4.13] remove shared VPC cred request
Backporting credential requests has the potential of breaking upgrades,
specifically if clusters are running in mint mode and have removed
the root credentials during z-stream upgrades.

This commit removes the 4.13 bump to the cred request for the AWS
shared VPC feature. The rest of the feature remains intact so that
it can be utilized by managed services which does not depend on
credential requests.