Remove dependency on canary route from various e2e tests
TestClientTLS, TestMTLSWithCRLs, and TestRouterCompressionOperation
queried the canary route to test various router config options. However,
with the canary being switched to use a passthrough route, many router
config options no longer apply. This commit changes each test to deploy
their own backend and route.

TestClientTLS and TestMTLSWithCRLs were both changed to use the echo pod
backend, which echoes the contents of the request back to the client.

TestRouterCompressionOperation requires that the response have a
specific content-type header, which the echo pod doesn't provide. It now
deploys a httpd backend, which does include the content-type header, and
has been updated to enable compression on the appropriate content-type.

This is part of the fix for OCPBUGS-9037