Allow setting router image in hack/release-local.sh
The default router image in the ingress operator deployment manifest is
not up to date, and sometimes causes red herring issues when debugging
changes.

With this change, a user can set the IMAGE environment variable to a
router image pullspec, and hack/release-local.sh will update the
deployment manifest to use the specified image.

Alternatively, a user can set the GET_IMAGE environment variable to a
non-empty value, and hack/release-local.sh will query the cluster in the
user's default kubeconfig to automatically get the router image
currently in use and use the same image in its generated manifests.
Note that GET_IMAGE requires read access to the
openshift-ingress-operator/ingress-operator deployment.