Add additional azure subnet permissions for private clusters
When installing a private cluster on azure, the customer supplies a
custom resource group for their existing vnet to which the ingress
operator (among others) must have access to. From within this custom
vnet, the operator must be able to read and join the various existing
subnets that the customer has supplied.

https://issues.redhat.com/browse/OCPBUGS-25193