TestHostNetworkPortBinding: Delete t.Parallel()
Remove the t.Parallel() call inside TestHostNetworkPortBinding.  The test
is listed in TestAll as a serial test, but the test itself calls
t.Parallel(), which causes it to run in parallel with other serial tests.
This likely causes occasional failures of the test in CI.

This commit fixes OCPBUGS-30091.

https://issues.redhat.com/browse/OCPBUGS-30091

* test/e2e/operator_test.go (TestHostNetworkPortBinding): Remove
t.Parallel().